refactor(timelapse): #236 h264/h265 merge 抽公共 codecMuxer 基类
把 h264_go_merge.go (1003 LOC) 和 h265_go_merge.go (982 LOC) 中逐字节相同的
box 骨架(writeMoov/writeMvhd/writeTrak/writeMdia/writeMinf/writeStbl)合并到
新建的 codec_merge.go 的 codecMuxer 类型。sample-entry box (avc1/hvc1) 和
config box (avcC/hvcC) 通过 codecMuxer 的 sampleEntryType/configBoxType 字段
参数化,writeSampleEntry 单一方法服务两种 codec。

同时把散落在两个文件里、其实已被跨文件复用的 codec-agnostic 辅助函数
(splitAnnexB / bitReader / removeEmulationPrevention / skipScalingLists)
集中到 codec_merge.go —— 之前 bitReader 定义在 h264 文件却被 h265 跨文件
引用,重命名/移动任一个都会打破另一个,很脆弱。

writeFtyp → writeCodecFtyp(brand),writeH264Mdat/writeH265Mdat →
writeCodecMdat(w, frames, ctx, isParamSet) —— 唯一的 codec 差异(param-set
谓词 + ftyp brand)参数化掉。

结果:

- h264_go_merge.go: 1003 → 369 LOC(满足 #236 ≤400)

- h265_go_merge.go: 982 → 414 LOC(h265 SPS 解析器本身比 h264 大,略超 400)

- codec_merge.go: 663 LOC(共享基类 + 辅助函数)

- 净减 ~440 行,消除 PR #92/#95 时代改一处要动两文件的重复税

行为零变化:box 字节序列完全一致(同样的 StartBox/Marshal/EndBox 序列、
同样的字段值)。timelapse 全套测试(TestH264/H265GoMerger_*、TestBuildHvcC_*、
TestBuildAvcC、TestParse*Dimensions、TestSplitAnnexB、TestAutoDetectMerger_*、
frame_extractor 测试)全绿。

验证:Docker VM go1.26 + gofumpt v1.11.0 全清;go build ./... 全过;
go test -race ./internal/{timelapse,mp4util,muxer,merge}/... 全绿
(timelapse 33s)。
