+++ internal/timelapse/h265_go_merge.go
@@ -9,11 +9,8 @@ package timelapse
 import (
 	"bytes"
 	"context"
-	"encoding/binary"
 	"fmt"
 	"os"
-	"path/filepath"
-	"sort"
 	"time"
 
 	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/model"
@@ -52,7 +49,7 @@ func (m *H265GoMerger) Tier() MergeTier {
 // outputPath with the given fps, and returns a MergeResult.
 func (m *H265GoMerger) Merge(ctx context.Context, framesDir, outputPath string, fps int) (*MergeResult, error) {
 	// List and sort frame files.
-	frames, err := listH265FrameFiles(framesDir)
+	frames, err := listCodecFrameFiles(framesDir, "h265")
 	if err != nil {
 		return &MergeResult{Tier: TierGo, Error: err.Error()}, err
 	}
@@ -168,13 +165,15 @@ func (m *H265GoMerger) Merge(ctx context.Context, framesDir, outputPath string, 
 	}
 
 	// First pass: calculate moov size by writing to a buffer.
-	muxer := &h265Muxer{
-		frameCount:  len(frames),
-		sampleSizes: sampleSizes,
-		width:       width,
-		height:      height,
-		hvcCData:    hvcCData,
-		sampleDur:   sampleDuration,
+	muxer := &codecMuxer{
+		frameCount:        len(frames),
+		sampleSizes:       sampleSizes,
+		width:             width,
+		height:            height,
+		decoderConfigData: hvcCData,
+		sampleDur:         sampleDuration,
+		sampleEntryType:   [4]byte{'h', 'v', 'c', '1'},
+		configBoxType:     [4]byte{'h', 'v', 'c', 'C'},
 	}
 
 	buf := &bytesWriter{}
@@ -202,7 +201,7 @@ func (m *H265GoMerger) Merge(ctx context.Context, framesDir, outputPath string, 
 	w := mp4.NewWriter(f)
 
 	// Write ftyp.
-	ftypSize, err := writeH265Ftyp(w)
+	ftypSize, err := writeCodecFtyp(w, [4]byte{'h', 'v', 'c', '1'})
 	if err != nil {
 		return &MergeResult{Tier: TierGo, Error: err.Error()},
 			fmt.Errorf("write ftyp: %w", err)
@@ -227,7 +226,7 @@ func (m *H265GoMerger) Merge(ctx context.Context, framesDir, outputPath string, 
 	}
 
 	// Write mdat box (strips VPS/SPS/PPS from samples — they are in hvcC only).
-	if err := writeH265Mdat(w, frames, ctx); err != nil {
+	if err := writeCodecMdat(w, frames, ctx, isH265ParamSet); err != nil {
 		f.Close()
 		os.Remove(outputPath)
 		return &MergeResult{Tier: TierGo, Error: err.Error()}, err
@@ -243,391 +242,7 @@ func (m *H265GoMerger) Merge(ctx context.Context, framesDir, outputPath string, 
 	}, nil
 }
 
-// --- H.265 MP4 Muxer ---
-
-// h265Muxer builds the moov box structure for an H.265 video track.
-type h265Muxer struct {
-	frameCount  int
-	sampleSizes []uint32
-	width       int
-	height      int
-	hvcCData    []byte
-	sampleDur   time.Duration
-	chunkOffset int64
-}
-
-func (m *h265Muxer) writeMoov(w *mp4.Writer, chunkOffset int64) error {
-	_, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("moov")})
-	if err != nil {
-		return err
-	}
-	if err := m.writeMvhd(w); err != nil {
-		return err
-	}
-	if err := m.writeTrak(w, chunkOffset); err != nil {
-		return err
-	}
-	_, err = w.EndBox()
-	return err
-}
-
-func (m *h265Muxer) writeMvhd(w *mp4.Writer) error {
-	bi, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("mvhd")})
-	if err != nil {
-		return err
-	}
-
-	duration := uint32(m.frameCount) * uint32(m.sampleDur.Milliseconds())
-
-	mvhd := &mp4.Mvhd{
-		Timescale:   1000,
-		DurationV0:  duration,
-		Rate:        0x00010000,
-		Volume:      0x0100,
-		NextTrackID: 2,
-		Matrix: [9]int32{
-			0x00010000, 0, 0,
-			0, 0x00010000, 0,
-			0, 0, 0x40000000,
-		},
-	}
-	if _, err := mp4.Marshal(w, mvhd, mp4.Context{}); err != nil {
-		return err
-	}
-	_, err = w.EndBox()
-	_ = bi
-	return err
-}
-
-func (m *h265Muxer) writeTrak(w *mp4.Writer, chunkOffset int64) error {
-	bi, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("trak")})
-	if err != nil {
-		return err
-	}
-
-	// tkhd
-	bi2, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("tkhd")})
-	if err != nil {
-		return err
-	}
-	duration := uint32(m.frameCount) * uint32(m.sampleDur.Milliseconds())
-	tkhd := &mp4.Tkhd{
-		TrackID:    1,
-		DurationV0: duration,
-		Width:      uint32(m.width) << 16,
-		Height:     uint32(m.height) << 16,
-		Matrix: [9]int32{
-			0x00010000, 0, 0,
-			0, 0x00010000, 0,
-			0, 0, 0x40000000,
-		},
-	}
-	if _, err := mp4.Marshal(w, tkhd, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi2
-
-	// mdia
-	if err := m.writeMdia(w, chunkOffset); err != nil {
-		return err
-	}
-
-	_, err = w.EndBox()
-	_ = bi
-	return err
-}
-
-func (m *h265Muxer) writeMdia(w *mp4.Writer, chunkOffset int64) error {
-	bi, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("mdia")})
-	if err != nil {
-		return err
-	}
-
-	duration := uint32(m.frameCount) * uint32(m.sampleDur.Milliseconds())
-
-	// mdhd
-	bi2, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("mdhd")})
-	if err != nil {
-		return err
-	}
-	mdhd := &mp4.Mdhd{
-		Timescale:  1000,
-		DurationV0: duration,
-		Language:   [3]byte{0x15, 0xC0, 0x00},
-	}
-	if _, err := mp4.Marshal(w, mdhd, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi2
-
-	// hdlr
-	bi3, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("hdlr")})
-	if err != nil {
-		return err
-	}
-	hdlr := &mp4.Hdlr{
-		HandlerType: [4]byte{'v', 'i', 'd', 'e'},
-		Name:        "VideoHandler\x00",
-	}
-	if _, err := mp4.Marshal(w, hdlr, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi3
-
-	// minf > stbl
-	if err := m.writeMinf(w, chunkOffset); err != nil {
-		return err
-	}
-
-	_, err = w.EndBox()
-	_ = bi
-	return err
-}
-
-func (m *h265Muxer) writeMinf(w *mp4.Writer, chunkOffset int64) error {
-	bi, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("minf")})
-	if err != nil {
-		return err
-	}
-
-	// vmhd
-	bi2, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("vmhd")})
-	if err != nil {
-		return err
-	}
-	if _, err := mp4.Marshal(w, &mp4.Vmhd{Graphicsmode: 0}, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi2
-
-	// dinf > dref > url
-	bi3, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("dinf")})
-	if err != nil {
-		return err
-	}
-	bi4, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("dref")})
-	if err != nil {
-		return err
-	}
-	if _, err := mp4.Marshal(w, &mp4.Dref{EntryCount: 1}, mp4.Context{}); err != nil {
-		return err
-	}
-	bi5, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("url ")})
-	if err != nil {
-		return err
-	}
-	if _, err := mp4.Marshal(w, &mp4.Url{Location: ""}, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi5
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi4
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi3
-
-	// stbl
-	if err := m.writeStbl(w, chunkOffset); err != nil {
-		return err
-	}
-
-	_, err = w.EndBox()
-	_ = bi
-	return err
-}
-
-func (m *h265Muxer) writeStbl(w *mp4.Writer, chunkOffset int64) error {
-	bi, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("stbl")})
-	if err != nil {
-		return err
-	}
-
-	n := m.frameCount
-
-	// stsd > hvc1 + hvcC
-	bi2, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("stsd")})
-	if err != nil {
-		return err
-	}
-	if _, err := mp4.Marshal(w, &mp4.Stsd{EntryCount: 1}, mp4.Context{}); err != nil {
-		return err
-	}
-	if err := m.writeHvc1SampleEntry(w); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi2
-
-	// stts
-	bi6, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("stts")})
-	if err != nil {
-		return err
-	}
-	sampleDelta := uint32(m.sampleDur.Milliseconds())
-	sttsEntries := make([]mp4.SttsEntry, n)
-	for i := range sttsEntries {
-		sttsEntries[i] = mp4.SttsEntry{
-			SampleCount: 1,
-			SampleDelta: sampleDelta,
-		}
-	}
-	if _, err := mp4.Marshal(w, &mp4.Stts{EntryCount: uint32(n), Entries: sttsEntries}, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi6
-
-	// stsc
-	bi7, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("stsc")})
-	if err != nil {
-		return err
-	}
-	stscEntries := []mp4.StscEntry{
-		{FirstChunk: 1, SamplesPerChunk: uint32(n), SampleDescriptionIndex: 1},
-	}
-	if n == 0 {
-		stscEntries = nil
-	}
-	if _, err := mp4.Marshal(w, &mp4.Stsc{EntryCount: uint32(len(stscEntries)), Entries: stscEntries}, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi7
-
-	// stsz
-	bi8, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("stsz")})
-	if err != nil {
-		return err
-	}
-	if _, err := mp4.Marshal(w, &mp4.Stsz{SampleSize: 0, SampleCount: uint32(n), EntrySize: m.sampleSizes}, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi8
-
-	// stco
-	bi9, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("stco")})
-	if err != nil {
-		return err
-	}
-	stco := &mp4.Stco{EntryCount: 0, ChunkOffset: nil}
-	if n > 0 {
-		stco.EntryCount = 1
-		stco.ChunkOffset = []uint32{uint32(chunkOffset)}
-	}
-	if _, err := mp4.Marshal(w, stco, mp4.Context{}); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi9
-
-	_, err = w.EndBox()
-	_ = bi
-	return err
-}
-
-// writeHvc1SampleEntry writes the hvc1 sample entry with the hvcC box inside stsd.
-func (m *h265Muxer) writeHvc1SampleEntry(w *mp4.Writer) error {
-	bi, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("hvc1")})
-	if err != nil {
-		return err
-	}
-
-	hvc1 := &mp4.VisualSampleEntry{
-		SampleEntry: mp4.SampleEntry{
-			AnyTypeBox:         mp4.AnyTypeBox{Type: mp4.StrToBoxType("hvc1")},
-			DataReferenceIndex: 1,
-		},
-		Width:           uint16(m.width),
-		Height:          uint16(m.height),
-		Horizresolution: 0x00480000,
-		Vertresolution:  0x00480000,
-		FrameCount:      1,
-		Depth:           0x0018,
-	}
-	if _, err := mp4.Marshal(w, hvc1, mp4.Context{}); err != nil {
-		return err
-	}
-
-	// hvcC box (HEVC decoder configuration data)
-	bi2, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("hvcC")})
-	if err != nil {
-		return err
-	}
-	if _, err := w.Write(m.hvcCData); err != nil {
-		return err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi2
-
-	if _, err := w.EndBox(); err != nil {
-		return err
-	}
-	_ = bi
-	return nil
-}
-
 // --- Helpers ---
-
-// writeH265Ftyp writes the ftyp box for an HEVC MP4 file and returns its total size.
-func writeH265Ftyp(w *mp4.Writer) (int64, error) {
-	start, _ := w.Seek(0, 1)
-	bi, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("ftyp")})
-	if err != nil {
-		return 0, err
-	}
-
-	ftyp := &mp4.Ftyp{
-		MajorBrand:   [4]byte{'i', 's', 'o', 'm'},
-		MinorVersion: 0,
-		CompatibleBrands: []mp4.CompatibleBrandElem{
-			{CompatibleBrand: [4]byte{'i', 's', 'o', 'm'}},
-			{CompatibleBrand: [4]byte{'i', 's', 'o', '2'}},
-			{CompatibleBrand: [4]byte{'m', 'p', '4', '1'}},
-			{CompatibleBrand: [4]byte{'h', 'v', 'c', '1'}},
-		},
-	}
-	if _, err := mp4.Marshal(w, ftyp, mp4.Context{}); err != nil {
-		return 0, err
-	}
-	if _, err := w.EndBox(); err != nil {
-		return 0, err
-	}
-	_ = bi
-
-	end, _ := w.Seek(0, 1)
-	return end - start, nil
-}
 
 // buildHvcC builds the HEVCDecoderConfigurationRecord bytes from raw VPS, SPS,
 // and PPS NAL units (without start codes). Format per ISO 14496-15.
@@ -667,76 +282,6 @@ func isH265ParamSet(nalu []byte) bool {
 	}
 	nalType := (nalu[0] >> 1) & 0x3F
 	return nalType == 32 || nalType == 33 || nalType == 34 // VPS, SPS, PPS
-}
-
-// writeH265Mdat writes the mdat box with H.265 frame data, stripping VPS/SPS/PPS
-// from each frame. Parameter sets belong in hvcC only — including them in sample
-// data causes MEDIA_ERR_SRC_NOT_SUPPORTED in browsers.
-func writeH265Mdat(w *mp4.Writer, framePaths []string, ctx context.Context) error {
-	// First pass: compute total mdat payload size (excluding param sets).
-	var totalPayloadSize uint32
-	for _, path := range framePaths {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-		data, err := os.ReadFile(path)
-		if err != nil {
-			return fmt.Errorf("read frame %s: %w", path, err)
-		}
-		for _, nalu := range splitAnnexB(data) {
-			if isH265ParamSet(nalu) {
-				continue
-			}
-			totalPayloadSize += 4 + uint32(len(nalu))
-		}
-	}
-
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	default:
-	}
-
-	mdatBoxSize := uint64(8 + totalPayloadSize)
-	_, err := w.StartBox(&mp4.BoxInfo{Type: mp4.StrToBoxType("mdat"), Size: mdatBoxSize})
-	if err != nil {
-		return fmt.Errorf("start mdat: %w", err)
-	}
-
-	// Write each frame as length-prefixed NALUs, skipping param sets.
-	for _, path := range framePaths {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-
-		data, err := os.ReadFile(path)
-		if err != nil {
-			return fmt.Errorf("read frame for mdat %s: %w", path, err)
-		}
-
-		for _, nalu := range splitAnnexB(data) {
-			if isH265ParamSet(nalu) {
-				continue
-			}
-			lenBytes := make([]byte, 4)
-			binary.BigEndian.PutUint32(lenBytes, uint32(len(nalu)))
-			if _, err := w.Write(lenBytes); err != nil {
-				return fmt.Errorf("write NALU length: %w", err)
-			}
-			if _, err := w.Write(nalu); err != nil {
-				return fmt.Errorf("write NALU data: %w", err)
-			}
-		}
-	}
-
-	if _, err := w.EndBox(); err != nil {
-		return fmt.Errorf("end mdat: %w", err)
-	}
-	return nil
 }
 
 // --- SPS parsing for H.265 ---
@@ -863,36 +408,6 @@ func parseH265Dimensions(sps []byte) (width, height int) {
 	}
 
 	return width, height
-}
-
-// removeEmulationPrevention removes H.264/H.265 emulation prevention bytes
-// (0x00 0x00 0x03 → 0x00 0x00) from RBSP data.
-func removeEmulationPrevention(data []byte) []byte {
-	result := make([]byte, 0, len(data))
-	i := 0
-	for i < len(data) {
-		if i+2 < len(data) && data[i] == 0 && data[i+1] == 0 && data[i+2] == 3 {
-			result = append(result, 0, 0)
-			i += 3
-		} else {
-			result = append(result, data[i])
-			i++
-		}
-	}
-	return result
-}
-
-// --- Frame listing ---
-
-// listH265FrameFiles returns a sorted list of .h265 frame files in the directory.
-func listH265FrameFiles(dir string) ([]string, error) {
-	matches, err := filepath.Glob(filepath.Join(dir, "frame_*.h265"))
-	if err != nil {
-		return nil, err
-	}
-	// Sort to ensure correct frame order.
-	sort.Strings(matches)
-	return matches, nil
 }
 
 // Ensure H265GoMerger satisfies the TimelapseMerger interface.
